requirements.txt
huggingface
kagglehub
transformers
torch
opencv-python
accelerate
pillow
pandas
hf_xet